Adding TextPublicController for index/show method to live in, updating templates with basic logic around edit/delete/create.
