+++ app/Http/Controllers/HomeController.php
@@ -21,12 +21,4 @@ class HomeController extends Controller
     {
         return view('home');
     }
-
-    /**
-     * @return \Illuminate\Http\Response
-     */
-    public function welcome()
-    {
-        return view('welcome');
-    }
 }
